requirements.txt
torch
torchvision
pytorch-lightning
numpy
pillow
pycocotools
PyYAML
